+++ Runtime/SQLiteConnectionExtensions.cs
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Gil Barbosa Reis
+ * Copyright (c) 2025 Gil Barbosa Reis
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -20,11 +20,13 @@
  * SOFTWARE.
  */
 using System;
+using System.Collections.Generic;
+using System.IO;
 using System.Runtime.InteropServices;
-#if UNITY_2018_1_OR_NEWER
+using SQLite.Csv;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
-#endif
+using UnityEngine;
 
 namespace SQLite
 {
@@ -40,13 +42,23 @@ namespace SQLite
             try
             {
                 var bytes = new byte[size];
-                Marshal.Copy(buffer, bytes, 0, (int) size);
+                Marshal.Copy(buffer, bytes, 0, (int)size);
                 return bytes;
             }
             finally
             {
                 SQLite3.Free(buffer);
             }
+        }
+
+        public static SQLiteAsset SerializeToAsset(this SQLiteConnection db, string schema = null, SQLiteOpenFlags openFlags = SQLiteOpenFlags.ReadOnly, bool storeDateTimeAsTicks = true, string streamingAssetsPath = null)
+        {
+            SQLiteAsset asset = ScriptableObject.CreateInstance<SQLiteAsset>();
+            asset.Bytes = db.Serialize(schema);
+            asset.OpenFlags = openFlags;
+            asset.StoreDateTimeAsTicks = storeDateTimeAsTicks;
+            asset.StreamingAssetsPath = streamingAssetsPath;
+            return asset;
         }
 
         public static SQLiteConnection Deserialize(this SQLiteConnection db, byte[] buffer, string schema = null, SQLite3.DeserializeFlags flags = SQLite3.DeserializeFlags.None)
@@ -64,7 +76,6 @@ namespace SQLite
             return db;
         }
 
-#if UNITY_2018_1_OR_NEWER
         public static SQLiteConnection Deserialize(this SQLiteConnection db, NativeArray<byte> buffer, string schema = null, SQLite3.DeserializeFlags flags = SQLite3.DeserializeFlags.None)
         {
             return Deserialize(db, buffer, buffer.Length, schema, flags);
@@ -83,9 +94,7 @@ namespace SQLite
             }
             return db;
         }
-#endif
 
-#if UNITY_2021_2_OR_NEWER
         public static SQLiteConnection Deserialize(this SQLiteConnection db, ReadOnlySpan<byte> buffer, string schema = null, SQLite3.DeserializeFlags flags = SQLite3.DeserializeFlags.None)
         {
             return Deserialize(db, buffer, buffer.Length, schema, flags);
@@ -107,6 +116,64 @@ namespace SQLite
             }
             return db;
         }
-#endif
+
+        /// <summary>
+        /// Import a CSV data stream into the table named <paramref name="tableName"/> inside the database.
+        /// The table will be created if it doesn't exist yet.
+        /// </summary>
+        /// <param name="db">Open database connection</param>
+        /// <param name="tableName">Name of the table that should be filled with data from the CSV data stream.</param>
+        /// <param name="csvStream">Data stream with CSV-formatted contents.</param>
+        /// <param name="separator">Separator used for parsing the CSV. Defaults to comma.</param>
+        /// <param name="maxFieldSize">Maximum field size allowed.</param>
+        /// <exception cref="ArgumentNullException">Thrown if any of <paramref name="db"/>, <paramref name="tableName"/> and <paramref name="csvStream"/> are null.</exception>
+        /// <exception cref="CsvException">Thrown if an error is found while parsing the CSV data.</exception>
+        public static void ImportCsvToTable(this SQLiteConnection db, string tableName, TextReader csvStream, CsvReader.SeparatorChar separator = CsvReader.SeparatorChar.Comma, int maxFieldSize = int.MaxValue)
+        {
+            if (db == null)
+            {
+                throw new ArgumentNullException(nameof(db));
+            }
+            if (string.IsNullOrWhiteSpace(tableName))
+            {
+                throw new ArgumentNullException(nameof(tableName));
+            }
+            if (csvStream == null)
+            {
+                throw new ArgumentNullException(nameof(csvStream));
+            }
+
+            var columns = new List<string>();
+            bool parsingHeader = true;
+            db.RunInTransaction(() =>
+            {
+                foreach (string field in CsvReader.ParseStream(csvStream, separator, maxFieldSize))
+                {
+                    if (field == null)  // newline
+                    {
+                        string joinedColumns = string.Join(", ", columns);
+                        if (parsingHeader)
+                        {
+                            db.Execute($"CREATE TABLE IF NOT EXISTS {tableName} ({joinedColumns})");
+                            parsingHeader = false;
+                        }
+                        else
+                        {
+                            db.Execute($"INSERT INTO {tableName} VALUES ({joinedColumns})");
+                        }
+                        columns.Clear();
+                    }
+                    else
+                    {
+                        if (parsingHeader && string.IsNullOrWhiteSpace(field))
+                        {
+                            throw new CsvException("Header cannot have empty column name.");
+                        }
+
+                        columns.Add(SQLiteConnection.Quote(field));
+                    }
+                }
+            });
+        }
     }
 }